Revert "[#45] Seperate real estate posts from general posts"
