+++ src/main/java/com/koreandubai/handubi/controller/PostController.java
@@ -1,24 +1,30 @@
 package com.koreandubai.handubi.controller;
 
-import com.koreandubai.handubi.controller.dto.*;
+import com.koreandubai.handubi.controller.dto.CreatePostRequestDto;
+import com.koreandubai.handubi.controller.dto.DetailedPost;
+import com.koreandubai.handubi.controller.dto.EditPostRequestDto;
+import com.koreandubai.handubi.controller.dto.GetUploadedImage;
 import com.koreandubai.handubi.global.common.CategoryType;
 import com.koreandubai.handubi.global.common.StatusEnum;
 import com.koreandubai.handubi.global.common.SubCategoryType;
 import com.koreandubai.handubi.global.common.SuccessResponse;
-import com.koreandubai.handubi.service.RealEstatePostService;
+import com.koreandubai.handubi.service.PostService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
+import java.io.IOException;
 import java.util.List;
 
 @RequiredArgsConstructor
 @RestController
-@RequestMapping("/real-estate")
-public class RealEstateController {
+@RequestMapping("/posts")
+public class PostController {
 
-    private final RealEstatePostService postService;
+    private final PostService postService;
+
 
     @GetMapping
     public SuccessResponse getAllPosts(@RequestParam(required = false, value = "category") CategoryType categoryType,
@@ -26,7 +32,7 @@ public class RealEstateController {
                                        @RequestParam(required = false, defaultValue = "0", value = "page") int pageNo,
                                        @RequestParam(required = false, defaultValue = "createdAt", value = "criteria") String criteria) {
 
-        List<DetailedRealEstatePost> products = postService.getPosts(categoryType, subCategory, pageNo, criteria);
+        List<DetailedPost> products = postService.getPosts(categoryType, subCategory, pageNo, criteria);
 
         return SuccessResponse.builder()
                 .status(StatusEnum.OK)
@@ -38,7 +44,7 @@ public class RealEstateController {
     @GetMapping("/{id}")
     public SuccessResponse getPost(@PathVariable("id") long postId) {
 
-        DetailedRealEstatePost post = postService.getSinglePost(postId);
+        DetailedPost post = postService.getSinglePost(postId);
 
         return SuccessResponse.builder()
                 .status(StatusEnum.OK)
@@ -50,7 +56,7 @@ public class RealEstateController {
     @PostMapping("/new/{category}")
     public SuccessResponse createPost(HttpServletRequest request,
                                       @PathVariable(value = "category") CategoryType categoryType,
-                                      @Valid @RequestBody CreateRealEstatePostRequestDto requestDto) {
+                                      @Valid @RequestBody CreatePostRequestDto requestDto) {
 
         postService.createPost(request, categoryType, requestDto);
 
@@ -74,7 +80,7 @@ public class RealEstateController {
     @PutMapping("/edit/{id}")
     public SuccessResponse editPost(HttpServletRequest request,
                                     @PathVariable("id") long postId,
-                                    @Valid @RequestBody EditRealEstatePostRequestDto requestDto) {
+                                    @Valid @RequestBody EditPostRequestDto requestDto) {
 
         postService.editPost(request, postId, requestDto);
 
@@ -88,12 +94,36 @@ public class RealEstateController {
     public SuccessResponse searchPostsByKeyword(@NotBlank @RequestParam String keyword,
                                                 @RequestParam(required = false, defaultValue = "0", value = "page") int pageNo,
                                                 @RequestParam(required = false, defaultValue = "created_at", value = "criteria") String criteria) {
-        List<DetailedRealEstatePost> posts = postService.searchPostsByKeyword(keyword, pageNo, criteria);
+        List<DetailedPost> posts = postService.searchPostsByKeyword(keyword, pageNo, criteria);
 
         return SuccessResponse.builder()
                 .status(StatusEnum.OK)
                 .message("Successfully search posts")
                 .data(posts)
+                .build();
+    }
+
+    @PostMapping("/images")
+    public SuccessResponse uploadImage(@RequestParam("file") MultipartFile file) throws IOException {
+
+        String url = postService.uploadImage(file);
+
+        return SuccessResponse.builder()
+                .status(StatusEnum.OK)
+                .data(url)
+                .message("Successfully upload image")
+                .build();
+    }
+
+    @GetMapping("/images/{imageName}")
+    public SuccessResponse getImage(@PathVariable String imageName) {
+
+        GetUploadedImage image = postService.getImage(imageName);
+
+        return SuccessResponse.builder()
+                .status(StatusEnum.OK)
+                .data(image)
+                .message("Successfully get image")
                 .build();
     }
 }
