--- src/main/java/com/koreandubai/handubi/global/common/ProductStatus.java
@@ -1,11 +1,0 @@
-package com.koreandubai.handubi.global.common;
-
-public enum ProductStatus {
-
-    // Real Estate
-    EMPTY,
-    OCCUPIED,
-    CONSTRUCTION_IN_PROGRESS,
-    PROCESS,
-    COMPLETE
-}
