--- src/main/java/com/koreandubai/handubi/global/common/ProductType.java
@@ -1,14 +1,0 @@
-package com.koreandubai.handubi.global.common;
-
-public enum ProductType {
-
-    // Real Estate
-    APARTMENT,
-    VILLA,
-    CONDO,
-    TOWNHOUSE,
-    PENTHOUSE,
-    COMMERCIAL,
-
-    ETC
-}
