--- src/main/java/com/koreandubai/handubi/global/common/State.java
@@ -1,12 +1,0 @@
-package com.koreandubai.handubi.global.common;
-
-public enum State {
-
-    ABU_DHABI,
-    AJMAN,
-    DUBAI,
-    FUJAIRAH,
-    RAS_AL_KHAIMAH,
-    SHARJAH,
-    UMM_AL_QUWAIN
-}
