+++ src/main/java/com/koreandubai/handubi/global/common/SubCategoryType.java
@@ -22,9 +22,7 @@ public enum SubCategoryType {
     INFO,
     BUY,
     HANDOVER,
-    URGENT_HANDOVER,
     RENT,
-    OFF_PLAN,
 
     // Second Hand
     ELECTRONIC,
@@ -48,7 +46,7 @@ public enum SubCategoryType {
 
     // child care
     EDUCATION,
-    ACADEMY,
+    ACADEMIY,
 
     // Life
     HOSPITAL,
