+++ src/main/java/com/koreandubai/handubi/repository/PostRepository.java
@@ -14,9 +14,9 @@ import java.util.Optional;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
 
-    Page<Post> findAllByCategoryAndPostStatus(CategoryType categoryType, PostStatus postStatus, Pageable pageable);
+    Page<Post> findAllByCategoryAndStatus(CategoryType categoryType, PostStatus status, Pageable pageable);
 
-    Page<Post> findAllByCategoryAndSubCategoryAndPostStatus(CategoryType categoryType, SubCategoryType subCategory, PostStatus postStatus, Pageable pageable);
+    Page<Post> findAllByCategoryAndSubCategoryAndStatus(CategoryType categoryType, SubCategoryType subCategory, PostStatus status, Pageable pageable);
 
     Optional<Post> getPostsById(Long postId);
